refactor(RecentProductLocalDataSource): 최근 본 상품 DAO에서 Data Source로 변경
