+++ app/src/main/java/woowacourse/shopping/data/respository/recentproduct/RecentProductRepositoryImpl.kt
@@ -1,25 +1,21 @@
 package woowacourse.shopping.data.respository.recentproduct
 
-import android.content.Context
-import woowacourse.shopping.data.database.RecentProductDao
 import woowacourse.shopping.data.model.RecentProductEntity
-import woowacourse.shopping.data.model.Server
+import woowacourse.shopping.data.respository.recentproduct.source.local.RecentProductLocalDataSource
 
 class RecentProductRepositoryImpl(
-    context: Context,
-    server: Server,
+    private val recentProductLocalDataSource: RecentProductLocalDataSource,
 ) : RecentProductRepository {
-    private val recentProductDao = RecentProductDao(context, server)
 
     override fun getRecentProducts(limit: Int): List<RecentProductEntity> {
-        return recentProductDao.getAll(limit)
+        return recentProductLocalDataSource.getAllRecentProducts(limit)
     }
 
     override fun deleteNotTodayRecentProducts(today: String) {
-        recentProductDao.deleteNotToday(today)
+        recentProductLocalDataSource.deleteNotToday(today)
     }
 
     override fun addCart(productId: Long) {
-        recentProductDao.insertRecentProduct(productId)
+        recentProductLocalDataSource.insertRecentProduct(productId)
     }
 }
